organize tester and re-use nice styles between test pages
