+++ experiments/tester/src/RNTester/RNTester.win32.tsx
@@ -4,16 +4,15 @@ import * as React from 'react';
 import { AppRegistry } from 'react-native';
 import { FabricTester, IFabricTesterProps } from './FabricTester';
 import { ThemeProvider } from '@uifabricshared/theming-react-native';
-import { customRegistry } from './CustomThemes';
+import { customRegistry } from './TestComponents/Theme/CustomThemes';
 
-const RNTesterApp: React.FunctionComponent<IFabricTesterProps> =
-(props) => {
-    return (
-      <ThemeProvider registry={customRegistry}>
-        <FabricTester {...props} />
-      </ThemeProvider>
-    );
-}
+const RNTesterApp: React.FunctionComponent<IFabricTesterProps> = props => {
+  return (
+    <ThemeProvider registry={customRegistry}>
+      <FabricTester {...props} />
+    </ThemeProvider>
+  );
+};
 
 AppRegistry.registerComponent('RNTesterApp', () => RNTesterApp);
