+++ experiments/tester/src/RNTester/TestComponents/FocusTrapZone/index.ts
@@ -1,0 +1,1 @@
+export * from './FocusTrapZoneTest';
