+++ experiments/tester/src/RNTester/TestComponents/Link/LinkTest.win32.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import * as ReactNative from 'react-native';
 import { Stack, Link } from '@fluentui/react-native';
-import { Square } from './Square';
-import { stackStyle } from '../TesterStyles';
+import { Square } from '../Common/Square';
+import { stackStyle } from '../Common/styles';
 
 export const LinkTest: React.FunctionComponent<{}> = () => {
   const doPress = (): void => {
