+++ experiments/tester/src/RNTester/TestComponents/Link/index.ts
@@ -1,0 +1,1 @@
+export * from './LinkTest.win32';
