+++ experiments/tester/src/RNTester/TestComponents/Persona/PersonaTest.tsx
@@ -2,18 +2,18 @@ import * as React from 'react';
 import { View } from 'react-native';
 import { Separator } from '@fluentui/react-native';
 import { TextWin32 } from '@office-iss/react-native-win32';
-import { styles } from './styles';
+import { commonTestStyles } from '../Common/styles';
 import { StandardUsage } from './StandardUsage';
 import { CustomizeUsage } from './CustomizeUsage';
 
 export const PersonaTest: React.FunctionComponent<{}> = () => {
   return (
     <View>
-      <TextWin32 style={styles.section}>Standard Usage</TextWin32>
+      <TextWin32 style={commonTestStyles.section}>Standard Usage</TextWin32>
       <Separator />
       <StandardUsage />
 
-      <TextWin32 style={styles.section}>Customize Usage</TextWin32>
+      <TextWin32 style={commonTestStyles.section}>Customize Usage</TextWin32>
       <Separator />
       <CustomizeUsage />
     </View>
