+++ experiments/tester/src/RNTester/TestComponents/Persona/index.ts
@@ -1,0 +1,1 @@
+export * from './PersonaTest';
