+++ experiments/tester/src/RNTester/TestComponents/PersonaCoin/index.ts
@@ -1,0 +1,1 @@
+export * from './PersonaCoinTest';
