+++ experiments/tester/src/RNTester/TestComponents/Pressable/index.ts
@@ -1,0 +1,1 @@
+export * from './PressableTest';
