+++ experiments/tester/src/RNTester/TestComponents/RadioGroup/index.ts
@@ -1,0 +1,1 @@
+export * from './RadioGroupTest';
