+++ experiments/tester/src/RNTester/TestComponents/Separator/SeparatorTest.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Button, Stack, Separator, Text } from '@fluentui/react-native';
-import { stackStyle, separatorStackStyle } from '../TesterStyles';
+import { stackStyle, separatorStackStyle } from '../Common/styles';
 
 const BlueSeparator = Separator.customize({ tokens: { color: 'blue' } });
 const RedSeparator = Separator.customize({ tokens: { color: 'red' } });
