+++ experiments/tester/src/RNTester/TestComponents/Separator/index.ts
@@ -1,0 +1,1 @@
+export * from './SeparatorTest';
