+++ experiments/tester/src/RNTester/TestComponents/Theme/index.ts
@@ -1,0 +1,1 @@
+export * from './ThemeTest';
