+++ experiments/tester/src/RNTester/TestComponents/index.ts
@@ -1,13 +1,14 @@
 import * as React from 'react';
-import { ButtonFocusTest } from './ButtonFocusTest';
-import { PressableTest } from './PressableTest';
-import { LinkTest } from './LinkTest.win32';
-import { SeparatorTest } from './SeparatorTest';
-import { ThemeTest } from './ThemeTest';
-import { FocusTrapTest } from './FocusTrapZoneTest';
-import { PersonaCoinTest } from './PersonaCoin/PersonaCoinTest';
-import { RadioGroupTest } from './RadioGroupTest';
-import { PersonaTest } from './Persona/PersonaTest';
+import { ButtonFocusTest } from './Button';
+import { CalloutTest } from './Callout';
+import { FocusTrapTest } from './FocusTrapZone';
+import { LinkTest } from './Link';
+import { PersonaTest } from './Persona';
+import { PersonaCoinTest } from './PersonaCoin';
+import { PressableTest } from './Pressable';
+import { RadioGroupTest } from './RadioGroup';
+import { SeparatorTest } from './Separator';
+import { ThemeTest } from './Theme';
 
 export type TestDescription = {
   name: string;
@@ -18,6 +19,10 @@ export const allTestComponents: TestDescription[] = [
   {
     name: 'Button Test',
     component: ButtonFocusTest
+  },
+  {
+    name: 'Callout Test',
+    component: CalloutTest
   },
   {
     name: 'Focus Trap Zone Test',
