--- experiments/tester/src/RNTester/TesterStyles.ts
@@ -1,15 +1,0 @@
-import { IStackProps } from '@fluentui/react-native';
-
-export const stackStyle: IStackProps['style'] = {
-  borderWidth: 1,
-  borderColor: '#bdbdbd',
-  paddingVertical: 8,
-  paddingHorizontal: 12,
-  margin: 8
-};
-
-export const separatorStackStyle: IStackProps['style'] = {
-  height: 200,
-  flexDirection: 'row',
-  justifyContent: 'space-evenly'
-};
